Fix oversliding slides reset
Instead of reseting the slide index at either end of the slider and only
a few slides backward or forward, reset the index immediately when the
user slides outside of the first original slides(the ones originaly in
the HTML) and always reset so that we land inside the original slides.

This fixes the problem where at certain settings you can slide so far
that the next few slides become invisible.